add exl2 seperation
